Add files via upload
The program uses a class calc to encapsulate the calculator functionality. The class contains private member variables to store the input values and the operation symbol. The calculator method handles user input, performs the arithmetic operations, and displays the result or error messages.

Run the Program: Start the program to open the console interface.

HOW TO USE;-

---> Enter Numbers: Input the first and second numbers when prompted.
---> Select Operation: Enter the arithmetic operation you want to perform (+, -, *, or /).
---> View Result: The application will display the result of the operation.
---> Error Handling: If an invalid operation is entered, the application prompts the user to enter a valid operation.

Example:-

Enter 1st no. -> 10
 Enter 2nd no. -> 5
 Enter operation you want to perform on these two numbers: +
 The sum of the x + y = 15